Printer: one node-label emitter for dot and GDL
dotPrinter.cpp and GDLPrinter.cpp each carried the same four-arm switch
deciding what text goes inside a node -- symbol name, bitvector constant,
or kind name.  GDL carried it a third time, spelled differently, for the
per-use constant nodes it duplicates: that arm tested GetType() ==
BOOLEAN_TYPE, which for a constant is exactly the TRUE/FALSE case the
switch's default already handled.

Fold both files into GraphPrinter.cpp with one printNodeLabel().

The walks themselves stay separate.  They look alike -- same memo on
GetNodeNum(), same recursion -- but GDL annotates nodes, suppresses edge
labels for commutative operators, gives each use of a constant its own
node, and does not recurse into constants.  Threading those through a
shared walk needs a callback per difference and makes the GDL side
harder to follow than it is now, so the shared file just puts them side
by side where the remaining overlap is visible.

No output change.
