+++ lib/Printer/CMakeLists.txt
@@ -23,8 +23,7 @@ include_directories(${PROJECT_BINARY_DIR}/src/AST/)
 
 add_library(printer OBJECT
     AssortedPrinters.cpp
-    dotPrinter.cpp
-    GDLPrinter.cpp
+    GraphPrinter.cpp
     LispPrinter.cpp
     PLPrinter.cpp
     SMTLIB1Printer.cpp
